+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 -* Add your own contribution below
 * Improved support for [Bitbucket Branch Source Plugin](https://wiki.jenkins-ci.org/display/JENKINS/Bitbucket+Branch+Source+Plugin) - bartoszj
-* Refactoring of `danger local` command - hanneskaeufler
+* Refactoring of `danger local` and `danger pr` commands - hanneskaeufler
 
 ## 4.2.1